Fix failed tests due to mocking problems.
